+++ Plutarch/Internal/PLam.hs
@@ -10,18 +10,18 @@ import Data.Kind (Type)
 import Data.Text qualified as Text
 import GHC.Stack (HasCallStack, callStack, withFrozenCallStack)
 import Plutarch.Internal (
+  Config (Tracing),
   PType,
   S,
   Term,
   pgetConfig,
   plam',
   punsafeConstantInternal,
-  tracingMode,
   (:-->),
   pattern DoTracingAndBinds,
  )
 import Plutarch.Internal.PrettyStack (prettyStack)
-import Plutarch.Internal.Trace (ptrace)
+import Plutarch.Internal.Trace (ptraceInfo)
 import PlutusCore qualified as PLC
 
 {- $plam
@@ -46,8 +46,11 @@ class PLamN (a :: Type) (b :: PType) (s :: S) | a -> b, s b -> a where
 instance {-# OVERLAPPABLE #-} a' ~ Term s a => PLamN a' a s where
   plam f =
     let cs = callStack
-     in plam' \x -> pgetConfig \c -> case tracingMode c of
-          DoTracingAndBinds -> ptrace (mkstring $ prettyStack "L" cs) $ f x
+     in plam' $ \x -> pgetConfig $ \case
+          -- Note: This works because at the moment, DoTracingAndBinds is the
+          -- most general tracing mode.
+          Tracing _ DoTracingAndBinds ->
+            ptraceInfo (mkstring $ prettyStack "L" cs) $ f x
           _ -> f x
 
 instance (a' ~ Term s a, PLamN b' b s) => PLamN (a' -> b') (a :--> b) s where