+++ Plutarch/Monadic.hs
@@ -3,7 +3,11 @@ module Plutarch.Monadic ((>>=), (>>), fail) where
 import Prelude hiding (fail, (>>), (>>=))
 
 import Data.String (fromString)
-import Plutarch (pgetConfig, tracingMode, pattern DetTracing)
+import Plutarch (
+  Config (Tracing),
+  TracingMode (DetTracing),
+  pgetConfig,
+ )
 import Plutarch.Prelude
 
 {- | Bind function used within do syntax.
@@ -50,7 +54,6 @@ Enables elegant usage of 'ptrace' and similar.
 @
 -}
 fail :: String -> Term s a
-fail s =
-  pgetConfig \c -> case tracingMode c of
-    DetTracing -> ptraceError "Pattern matching failure in QualifiedDo syntax"
-    _ -> ptraceError $ fromString s
+fail s = pgetConfig $ \case
+  Tracing _ DetTracing -> ptraceInfoError "Pattern matching failure in QualifiedDo syntax"
+  _ -> ptraceInfoError $ fromString s